requirements.txt
gensim==3.8.3
hydra-colorlog==1.0.0
hydra-core==1.1.0
numpy==1.19.5
omegaconf==2.1.0
pandas==1.2.1
pytorch-crf==0.7.2
pytorch-lightning==1.1.3
scikit-learn==0.24.1
scipy==1.6.0
simple-elmo==0.6.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
torch==1.7.1
torchaudio==0.7.2
TorchCRF==1.1.0
torchvision==0.8.2
tqdm==4.56.0
transformers==4.5.1
